Gallery of Photots
The code creates a gallery of photos with the featured photo, a button for the darkening of the featured photo, and a placeholder to show all the thumbnail photos of the gallery. If you want, you can freely redesign and add more photos to make it a gallery of your own.